+++ src/main/java/nextstep/subway/auth/ui/interceptor/authentication/UserDetailsService.java
@@ -1,7 +1,5 @@
 package nextstep.subway.auth.ui.interceptor.authentication;
 
-import nextstep.subway.member.domain.LoginMember;
-
 public interface UserDetailsService {
     LoginMember loadUserByUsername(String principal);
 }
