+++ src/main/java/nextstep/subway/member/application/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
 package nextstep.subway.member.application;
 
 import nextstep.subway.auth.ui.interceptor.authentication.UserDetailsService;
-import nextstep.subway.member.domain.LoginMember;
+import nextstep.subway.auth.ui.interceptor.authentication.LoginMember;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
